% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nbinom.R
\name{calculate_nbinom_density}
\alias{calculate_nbinom_density}
\title{Calculate densities of a negative binomial distribution}
\usage{
calculate_nbinom_density(size, prob, lower_quantile)
}
\arguments{
\item{size}{The size parameter of a negative binomial distribution}

\item{prob}{The prob parameter of a negative binomial distribution}

\item{lower_quantile}{Coverage of quantile of a distribution}
}
\value{
A data frame of the densities of the negative binomial distribution
}
\description{
Calculate densities of a negative binomial distribution
}
