% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cpp_nbinom.R
\name{calculate_nbinom_density_cpp}
\alias{calculate_nbinom_density_cpp}
\title{Calculate densities of a negative binomial distribution in C++}
\usage{
calculate_nbinom_density_cpp(size, prob, lower_quantile, step)
}
\arguments{
\item{size}{The size parameter of a negative binomial distribution}

\item{prob}{The prob parameter of a negative binomial distribution}

\item{lower_quantile}{Coverage of quantile of a distribution}

\item{step}{The step of vector of quantiles}
}
\value{
A data frame of the densities of the negative binomial distribution
}
\description{
Calculate densities of a negative binomial distribution in C++
}
