% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nbinom.R
\name{calculate_nbinom_mu_from_size_prob}
\alias{calculate_nbinom_mu_from_size_prob}
\title{Calculate the mu of a negative binomial distribution from its size and prob parameters}
\usage{
calculate_nbinom_mu_from_size_prob(size, prob)
}
\arguments{
\item{size}{The size parameter of a negative binomial distribution}

\item{prob}{The prob parameter of a negative binomial distribution}
}
\value{
The mu based on the input size and prob
}
\description{
Calculate the mu of a negative binomial distribution from its size and prob parameters
}
