% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nbinom.R
\name{calculate_nbinom_size_from_prob_mu}
\alias{calculate_nbinom_size_from_prob_mu}
\title{Calculate the size of a negative binomial distribution from its prob and mu parameters}
\usage{
calculate_nbinom_size_from_prob_mu(prob, mu)
}
\arguments{
\item{prob}{The prob parameter of a negative binomial distribution}

\item{mu}{The mu parameter of a negative binomial distribution}
}
\value{
The size based on the input prob and mu
}
\description{
Calculate the size of a negative binomial distribution from its prob and mu parameters
}
