% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nbinom.R
\name{draw_nbinom_density}
\alias{draw_nbinom_density}
\title{Draw a density plot of a negative binomial distribution}
\usage{
draw_nbinom_density(size, prob, lower_quantile)
}
\arguments{
\item{size}{The size parameter of a negative binomial distribution}

\item{prob}{The prob parameter of a negative binomial distribution}

\item{lower_quantile}{Coverage of quantile of a distribution}
}
\value{
A drawable object to pass to plot()
}
\description{
Draw a density plot of a negative binomial distribution
}
