% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nbinom.R
\name{get_default_nbinom_prob}
\alias{get_default_nbinom_prob}
\title{Get the default prob parameter of negative binomial distributions}
\usage{
get_default_nbinom_prob()
}
\value{
The default prob parameter
}
\description{
Get the default prob parameter of negative binomial distributions
}
