% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nbinom.R
\name{get_default_nbinom_size}
\alias{get_default_nbinom_size}
\title{Get the default size parameter of negative binomial distributions}
\usage{
get_default_nbinom_size()
}
\value{
The initial size parameter
}
\description{
Get the default size parameter of negative binomial distributions
}
