% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nbinom.R
\name{get_default_step}
\alias{get_default_step}
\title{Get the default step of vector of quantiles}
\usage{
get_default_step()
}
\value{
The default step of vector of quantiles
}
\description{
Get the default step of vector of quantiles
}
