% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{is_numeric}
\alias{is_numeric}
\title{Check a variable is a non-NA single-element number.}
\usage{
is_numeric(x)
}
\arguments{
\item{x}{a variable to check}
}
\value{
TRUE if x is a non-NA single-element number, FALSE otherwise.
}
\description{
Check a variable is a non-NA single-element number.
}
