% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nbinom.R
\name{NbinomDist}
\alias{NbinomDist}
\title{R6 Class representing a negative binomial distribution}
\description{
A negative binomial distribution has size and prob parameters.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NbinomDist$new()}}
\item \href{#method-set_size}{\code{NbinomDist$set_size()}}
\item \href{#method-set_prob}{\code{NbinomDist$set_prob()}}
\item \href{#method-set_mu}{\code{NbinomDist$set_mu()}}
\item \href{#method-get_size}{\code{NbinomDist$get_size()}}
\item \href{#method-get_prob}{\code{NbinomDist$get_prob()}}
\item \href{#method-get_mu}{\code{NbinomDist$get_mu()}}
\item \href{#method-reset}{\code{NbinomDist$reset()}}
\item \href{#method-draw}{\code{NbinomDist$draw()}}
\item \href{#method-get_dataframe}{\code{NbinomDist$get_dataframe()}}
\item \href{#method-clone}{\code{NbinomDist$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize and set default parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NbinomDist$new(size, prob)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{The size parameter which must be a positive number}

\item{\code{prob}}{The prob parameter which must be positive and lower than 1.0}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_size"></a>}}
\if{latex}{\out{\hypertarget{method-set_size}{}}}
\subsection{Method \code{set_size()}}{
Set the size parameter of this negative binomial distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NbinomDist$set_size(size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{The size parameter which must be a positive number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_prob"></a>}}
\if{latex}{\out{\hypertarget{method-set_prob}{}}}
\subsection{Method \code{set_prob()}}{
Set the prob parameter of this negative binomial distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NbinomDist$set_prob(prob)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prob}}{The prob parameter which must be positive and lower than 1.0}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_mu"></a>}}
\if{latex}{\out{\hypertarget{method-set_mu}{}}}
\subsection{Method \code{set_mu()}}{
Set the mu parameter of this negative binomial distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NbinomDist$set_mu(mu)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mu}}{The mu parameter which must be a positive number.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_size"></a>}}
\if{latex}{\out{\hypertarget{method-get_size}{}}}
\subsection{Method \code{get_size()}}{
Get the size parameter of this negative binomial distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NbinomDist$get_size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The size parameter
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_prob"></a>}}
\if{latex}{\out{\hypertarget{method-get_prob}{}}}
\subsection{Method \code{get_prob()}}{
Get the prob parameter of this negative binomial distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NbinomDist$get_prob()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The prob parameter
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_mu"></a>}}
\if{latex}{\out{\hypertarget{method-get_mu}{}}}
\subsection{Method \code{get_mu()}}{
Get the mu parameter of this negative binomial distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NbinomDist$get_mu()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The mu parameter. This is derived from the size and prob parameters and may slightly different from the value set in set_mu(mu)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Reset parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NbinomDist$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-draw"></a>}}
\if{latex}{\out{\hypertarget{method-draw}{}}}
\subsection{Method \code{draw()}}{
Draw a density plot of this negative binomial distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NbinomDist$draw(lower_quantile)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lower_quantile}}{Coverage of quantile of this distribution}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A drawable object to pass to plot()
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-get_dataframe}{}}}
\subsection{Method \code{get_dataframe()}}{
Get a data frame to plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NbinomDist$get_dataframe(lower_quantile)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lower_quantile}}{Coverage of quantile of this distribution}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame to plot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NbinomDist$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
