% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{negativeBinomialApp}
\alias{negativeBinomialApp}
\title{Launch a Shiny app}
\usage{
negativeBinomialApp()
}
\description{
This function launches an instance of the Shiny app to plot negative binomial distributions interactively
}
