% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nbinom.R
\name{read_config_parameters}
\alias{read_config_parameters}
\title{Read parameters from the config file}
\usage{
read_config_parameters()
}
\value{
A parameter set
}
\description{
Read parameters from the config file
}
